+++ packages/espressocash_backend/test/share_escrow/utils.dart
@@ -1,6 +1,25 @@
 import 'package:solana/dto.dart';
 import 'package:solana/solana.dart';
 
+class Accounts {
+  Accounts({
+    required this.escrowAccount,
+    required this.receiverAccount,
+    required this.senderTokenAccount,
+    required this.vaultTokenAccount,
+    required this.senderAccount,
+    required this.depositorAccount,
+    required this.receiverTokenAccount,
+  });
+  final Ed25519HDKeyPair escrowAccount;
+  final Ed25519HDKeyPair receiverAccount;
+  final Ed25519HDKeyPair senderAccount;
+  final Ed25519HDKeyPair depositorAccount;
+  final Ed25519HDPublicKey senderTokenAccount;
+  final Ed25519HDPublicKey receiverTokenAccount;
+  final Ed25519HDPublicKey vaultTokenAccount;
+}
+
 Future<Ed25519HDPublicKey> createAccount({
   required SolanaClient client,
   required Ed25519HDKeyPair owner,
